requirements.txt
streamlit
pandas
python-docx
fpdf
pymupdf
dateparser
